Extract Models::DataProperties.
- Add column annotations to Dnd5e::Condition.
